5.10

+++ osa5/bloglist-frontend/src/components/Blog.js
@@ -27,10 +27,9 @@ const Blog = ({ blog, likeBlog, deleteBlog}) => {
 
   const handleRemove = (event) => {
     event.preventDefault()
-
-
-    console.log("handleremove")
-    deleteBlog(blog.id)
+    if(window.confirm(`Remove blog ${blog.title} by ${blog.author}`)){
+      deleteBlog(blog.id)
+    }
   }
